VMM: More warnings.
svn:sync-xref-src-repo-rev: r57339

+++ src/VBox/VMM/VMMAll/PGMAllPool.cpp
@@ -1,4 +1,4 @@
-/* $Id: PGMAllPool.cpp 26274 2010-02-05 04:29:07Z knut.osmundsen@oracle.com $ */
+/* $Id: PGMAllPool.cpp 26279 2010-02-05 05:03:44Z knut.osmundsen@oracle.com $ */
 /** @file
  * PGM Shadow Page Pool.
  */
@@ -966,7 +966,7 @@ DECLINLINE(int) pgmPoolAccessHandlerSTOSD(PVM pVM, PPGMPOOL pPool, PPGMPOOLPAGE 
         pgmPoolMonitorChainChanging(pVCpu, pPool, pPage, GCPhysFault, (RTGCPTR)pu32, uIncrement);
 #endif
 #ifdef IN_RC
-        *(uint32_t *)pu32 = pRegFrame->eax;
+        *(uint32_t *)(uintptr_t)pu32 = pRegFrame->eax;
 #else
         PGMPhysSimpleWriteGCPhys(pVM, GCPhysFault, &pRegFrame->rax, uIncrement);
 #endif
